Fix notification date time
